# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  mentor_session_participant.CustomError:
    properties:
      error:
        type: string
    type: object
  mentor_session_participant.EnrollmentDetails:
    properties:
      mentor_session_participant_id:
        type: integer
      session_details:
        $ref: '#/definitions/mentor_session_participant.SessionDetails'
      status:
        type: string
    type: object
  mentor_session_participant.EnrollmentResponse:
    properties:
      enrollments:
        items:
          $ref: '#/definitions/mentor_session_participant.EnrollmentDetails'
        type: array
      message:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  mentor_session_participant.RegistrationResponse:
    properties:
      message:
        type: string
    type: object
  mentor_session_participant.SessionDetails:
    properties:
      image_url:
        type: string
      mentor_session_id:
        type: integer
      mentor_session_title:
        type: string
      schedule:
        type: string
      short_description:
        type: string
    type: object
...
  session_review.CustomError:
    properties:
      error:
        type: string
    type: object
  session_review.GetAllReviewsResponse:
    properties:
      message:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/session_review.ReviewDetail'
        type: array
      session_id:
        type: integer
      total:
        type: integer
    type: object
  session_review.ReviewDetail:
    properties:
      rating:
        type: integer
      review:
        type: string
      review_id:
        type: integer
      user_details:
        $ref: '#/definitions/session_review.UserDetails'
    type: object
  session_review.SessionReviewRequest:
    properties:
      rating:
        type: integer
      review:
        type: string
    type: object
  session_review.SessionReviewResponse:
    properties:
      message:
        type: string
    type: object
  session_review.UserDetails:
    properties:
      fullname:
        type: string
      image_url:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a Pijar API
  title: Pijar API
  version: "1.0"
paths:
...
  /reviews/{session_id}:
    get:
      description: Get session reviews
      parameters:
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pagesize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/session_review.GetAllReviewsResponse'
        "400":
          description: Session not found
          schema:
            $ref: '#/definitions/session_review.CustomError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/session_review.CustomError'
      tags:
      - Session Review
    post:
      consumes:
      - application/json
      description: Create session review
      parameters:
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      - description: Session Review
        in: body
        name: session_review
        required: true
        schema:
          $ref: '#/definitions/session_review.SessionReviewRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/session_review.SessionReviewResponse'
        "400":
          description: User not found
          schema:
            $ref: '#/definitions/session_review.CustomError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/session_review.CustomError'
      security:
      - Bearer: []
      summary: Create session review
      tags:
      - Session Review
...
  /sessions/{session_id}/enroll:
    post:
      description: Used for learner to join a mentor session
      parameters:
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mentor_session_participant.RegistrationResponse'
        "400":
          description: Invalid session ID
          schema:
            $ref: '#/definitions/mentor_session_participant.CustomError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/mentor_session_participant.CustomError'
      security:
      - Bearer: []
      summary: Used for learner to join a mentor session
      tags:
      - Session Enrollments
  /sessions/enrollments:
    get:
      description: Get learner enrollments
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/mentor_session_participant.EnrollmentResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/mentor_session_participant.CustomError'
      security:
      - Bearer: []
      summary: Get learner enrollments
      tags:
      - Session Enrollments
...
securityDefinitions:
  Bearer:
    description: '"Type ''Bearer TOKEN'' to correctly set the API Key"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"